Improved nav bar link logic

+++ app/javascript/components/layouts/Nav.jsx
@@ -1,29 +1,37 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext } from "react";
 import { Link } from "react-router-dom";
 import UserContext from "../context/UserContext";
 
 const Nav = () => {
   const [user, setUser] = useContext(UserContext);
 
-  return (
-    <div>
-      <Link to="/">HOME</Link>
-      <a href="/users/sign_in">Sign In</a>
-      <a href="/users/sign_up">Sign Up</a>
-      <a href="/users/sign_up">Log Out</a>
-      <a rel="nofollow" data-method="delete" href="/users/sign_out">
-        Logout
-      </a>
-      {user.current_user ? (
+  let jsxResult;
+
+  if (user.current_user) {
+    jsxResult = (
+      <div>
+        <Link to="/">HOME</Link>
+        <Link to="/users">Users Index</Link>
         <Link to={`/users/${user.current_user.id}`}>
           {user.current_user.email}
         </Link>
-      ) : (
-        ""
-      )}
-      <Link to="/users">Users Index</Link>
-    </div>
-  );
+        <a href="/users/edit">Edit Profile</a>
+        <a rel="nofollow" data-method="delete" href="/users/sign_out">
+          Logout
+        </a>
+      </div>
+    );
+  } else {
+    jsxResult = (
+      <div>
+        <Link to="/">HOME</Link>
+        <a href="/users/sign_in">Sign In</a>
+        <a href="/users/sign_up">Sign Up</a>
+      </div>
+    );
+  }
+
+  return jsxResult;
 };
 
 export default Nav;
